corrected book module
